backend/demo_try/requirements.txt
alembic==1.8.1
asgiref==3.4.1
Authlib==0.15.5
azure-common==1.1.28
azure-core==1.25.0
azure-identity==1.9.0
azure-mgmt-core==1.3.2
azure-mgmt-rdbms==10.1.0
azure-mgmt-resource==21.0.0
azure-mgmt-subscription==3.0.0
Babel==2.10.3
backports.zoneinfo==0.2.1
bcrypt==3.2.2
bidict==0.22.0
blinker==1.5
boto3==1.20.23
botocore==1.23.23
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.9
click==8.1.3
cryptography==3.4.8
Django==4.0
django-storages==1.13.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
dnspython==2.2.1
email-validator==1.2.1
eventlet==0.33.0
Flask==2.1.3
Flask-Babel==2.0.0
Flask-Compress==1.12
Flask-Gravatar==0.5.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==4.1.5
Flask-SocketIO==5.2.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
httpagentparser==1.9.3
idna==2.10
importlib-metadata==4.12.0
importlib-resources==5.9.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
ldap3==2.9.1
Mako==1.2.1
MarkupSafe==2.1.1
msal==1.18.0
msal-extensions==0.3.1
msrest==0.7.1
oauthlib==3.2.0
paramiko==2.11.0
passlib==1.7.4
pgadmin4==6.12
Pillow==9.2.0
portalocker==2.5.1
psutil==5.9.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.4.0
PyNaCl==1.5.0
pyotp==2.6.0
python-dateutil==2.8.2
python-engineio==4.3.4
python-socketio==5.7.1
pytz==2021.3
qrcode==7.3.1
requests==2.25.1
requests-oauthlib==1.3.1
s3transfer==0.5.0
simplejson==3.17.6
six==1.16.0
speaklater3==1.4
SQLAlchemy==1.4.40
sqlparse==0.4.2
sshtunnel==0.4.0
typing_extensions==4.3.0
ua-parser==0.15.0
urllib3==1.26.7
user-agents==2.2.0
Werkzeug==2.0.3
WTForms==3.0.1
zipp==3.8.1
